+++ meta_package_manager/managers/apm.py
@@ -20,7 +20,7 @@ from __future__ import annotations
 import json
 from typing import Iterator
 
-from click_extra.platform import LINUX, MACOS, WINDOWS
+from click_extra.platform import ANY_UNIX_BUT_MACOS, MACOS, WINDOWS, ANY_PLATFORM
 
 from ..base import Package, PackageManager
 from ..capabilities import search_capabilities, version_not_implemented
@@ -42,7 +42,7 @@ class APM(PackageManager):
 
     homepage_url = "https://atom.io/packages"
 
-    platforms = frozenset({LINUX, MACOS, WINDOWS})
+    platforms = ANY_PLATFORM
 
     requirement = "1.0.0"