+++ meta_package_manager/managers/cargo.py
@@ -20,7 +20,7 @@ from __future__ import annotations
 import re
 from typing import Iterator
 
-from click_extra.platform import LINUX, MACOS, WINDOWS
+from click_extra.platform import ANY_UNIX_BUT_MACOS, MACOS, WINDOWS, ANY_PLATFORM
 
 from ..base import Package, PackageManager
 from ..capabilities import search_capabilities, version_not_implemented
@@ -32,7 +32,7 @@ class Cargo(PackageManager):
 
     homepage_url = "https://doc.rust-lang.org/cargo/"
 
-    platforms = frozenset({MACOS, LINUX, WINDOWS})
+    platforms = ANY_PLATFORM
 
     requirement = "1.0.0"