+++ meta_package_manager/managers/dnf.py
@@ -20,7 +20,7 @@ from __future__ import annotations
 import re
 from typing import Iterator
 
-from click_extra.platform import LINUX
+from click_extra.platform import ANY_UNIX_BUT_MACOS
 
 from ..base import Package, PackageManager
 from ..capabilities import search_capabilities, version_not_implemented
@@ -35,7 +35,7 @@ class DNF(PackageManager):
 
     homepage_url = "https://github.com/rpm-software-management/dnf"
 
-    platforms = frozenset({LINUX})
+    platforms = ANY_UNIX_BUT_MACOS
 
     requirement = "4.0.0"