+++ meta_package_manager/managers/zypper.py
@@ -28,7 +28,7 @@ else:
     from typing import TypedDict
 
 import xmltodict
-from click_extra.platform import LINUX
+from click_extra.platform import ANY_UNIX_BUT_MACOS
 
 from ..base import Package, PackageManager
 from ..capabilities import version_not_implemented
@@ -47,7 +47,7 @@ class Zypper(PackageManager):
 
     homepage_url = "https://en.opensuse.org/Portal:Zypper"
 
-    platforms = frozenset({LINUX})
+    platforms = ANY_UNIX_BUT_MACOS
 
     requirement = "1.14.0"