+++ changelog.md
@@ -6,6 +6,8 @@
 This version is not released yet and is under active development.
 ```
 
+- \[mpm\] Unlock run on any Unix-like platform. Closes {issue}`872`.
+- \[mpm\] Activate `brew` on Windows Subsystem for Linux v2.
 - \[mpm\] Run tests on Python `3.12-dev`.
 
 ## {gh}`5.10.2 (2022-12-19) <compare/v5.10.1...v5.10.2>`